requirements.txt
torch>=1.0.0
torchvision
dominate>=2.3.1
dill
scikit-image
tensorboardX
matplotlib
Pillow
POT
tqdm
openpyxl
